backend: Dockerfile: fix entrypoint syntax
The shell form of the entrypoint instruction does not allow for
additional arguments/parameters to be supplied, which causes the Docker
Compose file to break as the command supplied in the Docker Compose file
is not executed. To fix this issue and allow for shell variable
substitution, invoke a shell directly in the exec form.

See more info here:
https://docs.docker.com/engine/reference/builder/#exec-form-entrypoint-example